Create Super Key
ID: Suganth-Mohan-201906-0323-0310
Description: Create a super key generically by combining the keys of MySQL table key columns with a delimiter and increase the speed significantly based on the total number of columns.
Language: Python
Comments: Cannot contain separator/delimiter in the record, even if it does, it will work for compare but becomes unusable after that.