Execute Bash Commands Securily
ID: Suganth-Mohan-201906-1517-3939
Description: Execute Bash Commands by Reading the STDOUT and STDERR to see if there has been any errors occured.
Language: Python 3.5
Comments: The Constraint is that if the command prints output in STDERR it will also be considered as an error and needs to be set an execption for it.
Co-Authored-By: kasiramakrishnan <kasiramakrishnan@users.noreply.github.com>